adding support for json_line input format
Now logle can parse json_line files and is not limited by the size of input
file, bacause input can be read line by line.

Now plaso_analyzer is initialized not with a filename, but with a Subclass of
JsonDocumentIterator. FullJson for old JSON format and StreamJson for json_line
format.